get tests working again

+++ hybrid/AppDelegate.swift
@@ -23,6 +23,12 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         return true
     }
     
+    static var runningInTests:Bool {
+        get {
+            return NSClassFromString("XCTest") != nil
+        }
+    }
+    
     func application(application: UIApplication, didFinishLaunchingWithOptions launchOptions: [NSObject: AnyObject]?) -> Bool {
         
        
@@ -63,15 +69,20 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             
             let rootController = HybridNavigationController.create()
             
-            // todo: remove
-            ServiceWorkerManager.clearActiveServiceWorkers()
-            try Db.mainDatabase.inDatabase({ (db) in
-                db.executeUpdate("DELETE FROM service_workers", withArgumentsInArray: nil)
-            })
-//
-//            rootController.pushNewHybridWebViewControllerFor(NSURL(string:"https://www.gdnmobilelab.com/app-demo")!)
+            if AppDelegate.runningInTests == false {
+                // todo: remove
+                ServiceWorkerManager.clearActiveServiceWorkers()
+                try Db.mainDatabase.inDatabase({ (db) in
+                    db.executeUpdate("DELETE FROM service_workers", withArgumentsInArray: nil)
+                })
+
+                // rootController.pushNewHybridWebViewControllerFor(NSURL(string:"https://www.gdnmobilelab.com/app-demo")!)
+                
+                rootController.pushNewHybridWebViewControllerFor(NSURL(string:"https://eac8a863.ngrok.io/app-demo")!)
+            }
             
-            rootController.pushNewHybridWebViewControllerFor(NSURL(string:"https://eac8a863.ngrok.io/app-demo")!)
+            
+           
         
             AppDelegate.window!.rootViewController = rootController
